ci: Use wdio options to inject params. Move GitHubService to ci.mjs.
