+++ test/ci.mjs
@@ -70,6 +70,7 @@ export class GitHubStatus {
 
 /**
  * build returns an identifier for the build in question.
+ * @returns {String}
  */
 export async function build() {
   if (process.env.HEAD_BRANCH) {
@@ -92,5 +93,105 @@ export function gitBuild() {
       }
       resolve(stdout.trim());
     });
+  });
+}
+
+export class GitHubService {
+  constructor(opts, caps, config) {
+    this.gh = new GitHubStatus({
+      context: opts.context,
+      sha: opts.sha,
+      repo: opts.repo,
+      token: opts.token,
+    });
+    opts.links?.forEach(l => l.then(({priority, url}) => this.gh.setTarget(priority, url)));
+    this.specs = new Set();
+  }
+
+  onPrepare(config, caps) {
+    this.counts = {
+      caps: caps.length,
+      specs: 0,
+      total: 0,
+      running: 0,
+      finished: 0,
+      passed: 0,
+      failed: 0,
+    };
+    this.#buildPrivateURL();
+    this.#findPublicURL();
+  }
+
+  onWorkerStart(cid, cap, specs, args, execArgv) {
+    for (let spec of specs) {
+      this.specs.add(spec);
+    }
+    this.counts.specs = this.specs.size;
+    this.counts.total = this.counts.caps * this.counts.specs;
+    this.counts.running++;
+    this.#update('pending');
+  }
+
+  onWorkerEnd(cid, exitCode, specs, retries) {
+    this.counts.running--;
+    this.counts.finished++;
+    if (exitCode == 0) {
+      this.counts.passed++;
+    } else {
+      this.counts.failed++;
+    }
+    this.#update('pending');
+  }
+
+  onComplete(exitCode, config, capabilities, results) {
+    this.counts = {...this.counts, ...results};
+    return this.#update(exitCode ? 'failure' : 'success');
+  }
+
+  #update(state) {
+    const c = this.counts;
+    const desc = state === 'pending'
+      ? `${c.running} running with ${c.passed} passed & ${c.failed} failed of ${c.total} total (${(100*c.finished/c.total).toFixed(0)}%)`
+      : `${c.passed} passed & ${c.failed} failed`;
+    return this.gh.update({
+      state: state,
+      description: desc,
+    });
+  }
+}
+
+export function browserstackPrivateURL(buildName) {
+  return {
+    priority: 0,
+    url: 'https://automate.browserstack.com/dashboard/v2/search?type=builds&query=' + encodeURIComponent(buildName),
+  }
+}
+
+/**
+ * Poll the browserstack API to find the public URL of our build and give it to
+ * the github client for inclusion on commit statuses, overriding any URL
+ * generated by _ghInitPvt.
+ */
+export function browserstackPublicURL(buildName) {
+  return new Promise(function(resolve) {
+    setTimeout(findBuild, 2000);
+    function findBuild() {
+      fetch(
+        'https://api.browserstack.com/automate/builds.json',
+        {headers: {'Authorization': 'Basic '+Buffer.from(user + ':' + key).toString('base64')}}
+      )
+      .then(res => {
+        if (!res.ok) throw new Error(res.url + ' returned status ' + res.statusCode);
+        return res.json();
+      })
+      .then(builds => Array.isArray(builds) && builds.filter(b => b.automation_build && b.automation_build.name == buildName).pop())
+      .then(build => build
+        ? resolve({priority: 1, url: build.automation_build.public_url})
+        : setTimeout(findBuild, 1000))
+      .catch(err => {
+        console.error('Error fetching browserstack builds list', err);
+        setTimeout(findBuild, 1000);
+      });
+    }
   });
 }
